+++ frontend/src/components/HomePage/HomePage.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import AppBar from '@material-ui/core/AppBar';
 import Button from '@material-ui/core/Button';
+import ButtonGroup from '@material-ui/core/ButtonGroup';
 import SvgIcon from "@material-ui/core/SvgIcon";
 import Card from '@material-ui/core/Card';
 import CardActions from '@material-ui/core/CardActions';
@@ -12,9 +13,8 @@ import Box from '@material-ui/core/Box';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
-import { styled, alpha } from '@material-ui/core/styles';
-import InputBase from '@material-ui/core/InputBase';
 import SearchIcon from '@material-ui/icons/Search';
+import {Search, SearchIconWrapper, StyledInputBase, ButtonBox} from './HomePageStyles';
 
 function HomeIcon(props) {
   return (
@@ -22,55 +22,13 @@ function HomeIcon(props) {
       <path d="M10 20v-6h4v6h5v-8h3L12 3 2 12h3v8z" />
     </SvgIcon>
   );
-}
-
-const Search = styled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginLeft: 0,
-  width: '100%',
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: theme.spacing(1),
-    width: 'auto',
-  },
-}));
-
-const SearchIconWrapper = styled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: '12ch',
-      '&:focus': {
-        width: '20ch',
-      },
-    },
-  },
-}));
+};
 
 
 const cards = [1, 2, 3, 4];
 
 
-function Community() {
+function HomePage() {
 
   return (
     <React.Fragment>
@@ -90,6 +48,12 @@ function Community() {
               inputProps={{ 'aria-label': 'search' }}
             />
           </Search>
+          <ButtonBox>
+            <ButtonGroup variant="contained" aria-label="outlined primary button group">
+              <Button style={{textTransform: 'none'}} onClick = {() => {alert("log in")}}>Log In</Button>
+              <Button style={{textTransform: 'none'}} onClick = {() => {alert("sign up")}}>Sign Up</Button>
+             </ButtonGroup>
+         </ButtonBox>
         </Toolbar>
       </AppBar>
       <main>
@@ -182,4 +146,4 @@ function Community() {
 }
 
 
-export default Community;
+export default HomePage;
