+++ frontend/src/components/HomePage/HomePageStyles.js
@@ -1,0 +1,54 @@
+import { styled, alpha } from '@material-ui/core/styles';
+import InputBase from '@material-ui/core/InputBase';
+import Box from '@material-ui/core/Box';
+const Search = styled('div')(({ theme }) => ({
+    position: 'relative',
+    left: '60%',
+    borderRadius: theme.shape.borderRadius,
+    backgroundColor: alpha(theme.palette.common.white, 0.15),
+    '&:hover': {
+      backgroundColor: alpha(theme.palette.common.white, 0.25),
+    },
+    marginLeft: 0,
+    width: '100%',
+    [theme.breakpoints.up('sm')]: {
+      marginLeft: theme.spacing(1),
+      width: 'auto',
+    },
+  }));
+  
+  const SearchIconWrapper = styled('div')(({ theme }) => ({
+    padding: theme.spacing(0, 2),
+    height: '100%',
+    position: 'absolute',
+    pointerEvents: 'none',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  }));
+  
+  const StyledInputBase = styled(InputBase)(({ theme }) => ({
+    color: 'inherit',
+    '& .MuiInputBase-input': {
+      padding: theme.spacing(1, 1, 1, 0),
+      // vertical padding + font size from searchIcon
+      paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+      transition: theme.transitions.create('width'),
+      width: '100%',
+      [theme.breakpoints.up('sm')]: {
+        width: '12ch',
+        '&:focus': {
+          width: '20ch',
+        },
+      },
+    },
+  }));
+
+  const ButtonBox = styled(Box)(({theme}) => ({
+    position: 'relative',
+    left: '65%',
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center','& > *': { m: 1,},
+  }))
+  export {Search, SearchIconWrapper, StyledInputBase, ButtonBox};
